fix(#4787): relocate palette.py outside textual_chat's TTY-only import boundary
Found implementing #4787①'s constant migration, verified empirically
before shipping anything: textual_chat/__init__.py's own docstring
states "must only ever be imported on the TTY path" — it eagerly
imports textual/textual_flowview through its submodules (.app/.chrome/
.completion/.gutter/.intervention_panel/.presenter). Python always runs
a package's __init__.py before any of its submodules, so ANY import of
textual_chat.palette — even a direct one, even though palette.py itself
has zero framework imports — paid that cost regardless. Confirmed
directly: `'textual' in sys.modules` was False after importing
interfaces.repl.renderer alone, True the moment textual_chat.palette
was imported.

This blocked renderer.py (the plain/--cui/non-TTY chat path, which must
stay usable without flowview installed) from ever reading these tokens
without breaking that invariant — directly informing #4840's still-open
REPL mechanism question.

Lead-coder's ruling (structural call): relocate to
src/reyn/interfaces/palette.py (verified: interfaces/__init__.py is
docstring-only, zero imports) rather than lazy-importing inside
textual_chat's own __init__ — moving palette out is the narrower
change, and the eager import there is an intentional invariant, not a
bug to route around.

All 14 real consumers updated to import from the new location directly
— no re-export shim left behind, per lead-coder's explicit instruction.
Found via `git grep` for every import FORM (package-level, relative,
function-local, multi-name-on-one-line) rather than trusting a single
pattern — the 14th (test_activity_row_3693.py's
`from ... import TextualChatApp, palette`) only surfaced by running the
actual test suite, not by grep alone.

interfaces/inline/textual_chat/:
  activity_row.py, app.py, completion.py, intervention_panel.py,
  presenter.py, rewind_picker.py, search_bar.py, sent_queue.py,
  text_effect.py

tests/interfaces/:
  test_4691_phase_b_group_construction.py, test_activity_row_3693.py,
  test_recede_dim_3523.py, test_selection_not_bright_3542.py,
  test_tui_colour_tokens.py (the gate's own import, per lead-coder's
  explicit instruction — a gate that doesn't look at the actual new
  location guards nothing)

Falsified: confirmed the OLD import path worked before the move
(git stash), confirmed renderer.py alone still pulls in zero Textual
modules after the move, confirmed the NEW location is import-cheap
(zero Textual modules pulled in either).

Time-dependency check (#4845/#4847): 0 instances — no test in this
diff writes a time bound in either direction.
