+++ src/reyn/interfaces/inline/textual_chat/intervention_panel.py
@@ -108,7 +108,7 @@ from textual.widgets import (
     Tabs,
 )
 
-from reyn.interfaces.inline.textual_chat import palette
+from reyn.interfaces import palette
 
 from .presenter import _neutralized_label
